continued the int -> unsigned short int for port address what Alex started (Yeah, konquy shows html-pages again :-)
svn path=/trunk/kdebase/kioslave/; revision=24155

+++ gopher/gopher.cc
@@ -157,7 +157,7 @@ bool GopherProtocol::gopher_open( KURL &_url )
   // is a port specified in /etc/services, and if so use that
   // otherwise as a last resort use the "official" port of 110.
 
-  unsigned int port;
+  unsigned short int port;
   struct sockaddr_in server_name;
   memset(&server_name, 0, sizeof(server_name));
   static char buf[512];
